requirements.txt
fastapi
uvicorn
lunisolar-upgraded==0.1.4
